updates to rawinput based on Zak's comments
